--- services/web/server/src/simcore_service_webserver/exporter/formatters/__init__.py
@@ -1,4 +1,0 @@
-from ..exceptions import ExporterException
-from .formatter_v2 import FormatterV2
-
-__all__: tuple[str, ...] = ("ExporterException", "FormatterV2")
