--- services/web/server/src/simcore_service_webserver/exporter/formatters/sds/xlsx/__init__.py
@@ -1,3 +1,0 @@
-from .writer import write_xlsx_files
-
-__all__ = ["write_xlsx_files"]
