+++ services/web/server/src/simcore_service_webserver/exporter/sds/xlsx/templates/utils.py
